Separate solver types into direct imports
