+++ momentGW/thc.py
@@ -6,7 +6,8 @@ import h5py
 import numpy as np
 from scipy.special import binom
 
-from momentGW import init_logging, ints, logging, tda, util
+from momentGW import init_logging, ints, logging, util
+from momentGW.tda import dTDA as DFdTDA
 
 
 class Integrals(ints.Integrals):
@@ -244,7 +245,7 @@ class Integrals(ints.Integrals):
         return self._blocks["La"]
 
 
-class dTDA(tda.dTDA):
+class dTDA(DFdTDA):
     """
     Compute the self-energy moments using dTDA and numerical integration
     with tensor-hypercontraction.
