+++ momentGW/uhf/evgw.py
@@ -6,7 +6,8 @@ constraints for molecular systems.
 import numpy as np
 
 from momentGW import logging, util
-from momentGW import evGW, UGW
+from momentGW import evGW
+from momentGW.uhf import UGW
 
 
 class evUGW(UGW, evGW):
